Refactor: API요청은 클라이언트 사이드에서 하도록 변경
