+++ app/(nav)/myPage/page.tsx
@@ -1,112 +1,51 @@
-'use client';
+import { redirect } from 'next/navigation';
+import { getServerSession } from 'next-auth/next';
+import { authOptions } from '@/app/api/auth/[...nextauth]/route';
+import MyPageClient from '@/app/(nav)/myPage/MyPageClient';
+import { API_URL_CONFIG } from '@/config/apiEndPointConfig';
+import axiosInstance from '@/lib/axios';
 
-import { useState, useEffect } from 'react';
-import { useRouter } from 'next/navigation';
-import { signOut, useSession } from 'next-auth/react';
-import TopBar from '@/components/Common/TopBar/TopBar';
-import {
-  MyPageWrapper,
-  ProfileWrapper,
-  SmallButtonWrapper,
-  Separator,
-} from '@/app/(nav)/myPage/myPage.styles';
-import ProfileInfo from '@/components/Layout/MyPage/ProfileInfo';
-import SettingButton from '@/components/Layout/MyPage/SettingButton';
-import ProfileButton from '@/components/Layout/MyPage/ProfileButton';
-import AlertModal from '@/components/Common/AlertModal/AlertModal';
+export default async function MyPage() {
+  // 서버에서 세션 정보를 가져옴
+  const session = await getServerSession(authOptions);
+  if (!session) {
+    console.error('세션이 없습니다. 로그인 페이지로 리다이렉트합니다.');
+    redirect('/login');
+  }
 
-const MyPage = () => {
-  const router = useRouter();
-  const [isLogoutModalOpen, setIsLogoutModalOpen] = useState(false);
-
-  // NextAuth 세션 훅
-  const { data: session } = useSession();
   // 세션에서 userId 추출
-  const userId = session?.user?.userId;
+  const userId = session.user?.userId;
+  if (!userId) {
+    console.error(
+      '세션에 userId가 없습니다. 로그인 페이지로 리다이렉트합니다.'
+    );
+    redirect('/login');
+  }
 
-  const [userData, setUserData] = useState({
-    email: '',
-    nickname: '',
-    activeCharacter: '',
-    level: 0,
-    exp: 0,
-  });
+  try {
+    console.log(`서버 컴포넌트에서 userId: ${userId}로 직접 백엔드 API 호출`);
+    console.log(`호출 URL: ${API_URL_CONFIG.USER_SERVICE.USERS}/${userId}`);
 
-  useEffect(() => {
-    const fetchUserData = async () => {
-      if (!userId) {
-        return;
-      }
+    // 라우터를 거치지 않고 직접 백엔드 API 호출
+    const response = await axiosInstance.get(
+      `${API_URL_CONFIG.USER_SERVICE.USERS}/${userId}`
+    );
 
-      try {
-        const response = await fetch(`/api/userInfo`, {
-          headers: {
-            'User-Id': String(userId), // 문자열 변환
-          },
-        });
+    const { email, nickname, activeCharacter, exp, level } = response.data;
+    console.log('서버에서 받아온 userData:', response.data);
 
-        if (!response.ok) {
-          throw new Error('유저 데이터를 불러오는데 실패했습니다.');
-        }
-
-        const data = await response.json();
-        setUserData({
-          email: data.email,
-          nickname: data.nickname,
-          activeCharacter: data.activeCharacter,
-          level: data.level,
-          exp: data.exp,
-        });
-      } catch (error) {
-        console.error('유저 데이터 fetch 에러:', error);
-      }
+    const userData = {
+      email,
+      nickname,
+      activeCharacter,
+      level,
+      exp,
     };
 
-    fetchUserData();
-  }, [userId]);
-
-  const handleLogout = async () => {
-    setIsLogoutModalOpen(false);
-    await signOut({ callbackUrl: '/' });
-  };
-
-  return (
-    <>
-      <TopBar NavType="default" label="마이페이지" />
-      <MyPageWrapper>
-        <ProfileWrapper>
-          <ProfileInfo
-            email={userData.email}
-            nickname={userData.nickname}
-            activeCharacter={userData.activeCharacter}
-          />
-          <SmallButtonWrapper>
-            <ProfileButton
-              onClick={() => router.push('/myPage/nicknameChange')}
-            >
-              닉네임 변경
-            </ProfileButton>
-          </SmallButtonWrapper>
-        </ProfileWrapper>
-        <Separator />
-        <SettingButton
-          label="로그아웃"
-          onClick={() => setIsLogoutModalOpen(true)}
-        />
-        <SettingButton
-          label="회원탈퇴"
-          onClick={() => router.push('/myPage/accountCencellation')}
-        />
-      </MyPageWrapper>
-
-      <AlertModal
-        isOpen={isLogoutModalOpen}
-        onClose={() => setIsLogoutModalOpen(false)}
-        onConfirm={handleLogout}
-        title="로그아웃 하시겠습니까?"
-      />
-    </>
-  );
-};
-
-export default MyPage;
+    // 클라이언트 컴포넌트로 데이터를 전달
+    return <MyPageClient userData={userData} />;
+  } catch (error) {
+    console.error('서버사이드 에러:', error);
+    redirect('/error');
+  }
+}
